// src/app/interfaces/trail.ts
export interface Trail {
    ascent: number;
    conditionDate: Date;
    conditionDetails: any;
    conditionStatus: string;
    descent: number;
    difficulty: string;
    high: number;
    id: number;
    imgMedium: string;
    imgSmall: string;
    imgSmallMed: string;
    imgSqSmall: string;
    latitude: number;
    length: number;
    location: string;
    longitude: number;
    low: number;
    name: string;
    starVotes: number;
    stars: number;
    summary: string;
    type: string;
    url: string;
}